private_rest_api/private_rest_api/requirements.txt
# Django packages:
django
djangorestframework
django-filter
django-celery-beat

# Process Scheudler packages:
Redis
celery

# Database packages:
psycopg2

# Additional packages:
markdown
praw
